scripteja tehty

+++ poista-a-record.sh
@@ -3,14 +3,14 @@
 set -euo pipefail
 
 laitaPorttiTakaisin() {
-    portit="$1"
-    nginx_conf="$2"
-    domain=${nginx_conf##*/}
-    domain=${domain%.*}
+    local portit="$1"
+    local nginx_conf="$2"
+    local domain=${nginx_conf##*/}
+    local domain=${domain%.*}
 
     local portti=$(grep -P "proxy_pass http://$domain:\d{4};" $nginx_conf || :)
-    portti=${portti##*:}
-    portti=${portti%;}
+    local portti=${portti##*:}
+    local portti=${portti%;}
 
     if [[ -z $portti ]]; then
 	echo "Portti on tyhjä merkkijono!"
@@ -21,7 +21,7 @@ laitaPorttiTakaisin() {
 
 while getopts "h" flag; do
     case "${flag}" in
-        h) echo "Käyttö: poista-a-record domain domain-id a-record" 
+        h) echo "Käyttö: poista-a-record domain a-record" 
 	   echo
 	   echo "Poistaa a-recordin Linodesta ja siihen liittyvät kansiot."
 	   echo
@@ -32,11 +32,10 @@ while getopts "h" flag; do
 done
 
 domain="$1"
-domain_id="$2"
-record="$3"
+record="$2"
 
-if [[ "$domain_id" =~ [a-zA-Z] ]]; then
-    echo "'domain-id' pitää olla kokonaisluku! ($domain_id)"
+if [[ ! "$record" =~ "$domain" ]]; then
+    echo "Anna record muodossa (record.)domain"
     exit 1
 fi
 
@@ -54,12 +53,28 @@ rm -rf "$log/$domain/$record"
 rm -rf "$nginx_conf"
 rm -rf "$nginx_conf"
 
-[[ "$record" =~ "$domain" && "$record" != "$domain" ]] \
-    && record="${record%.*}"
+[[ "$record" != "$domain" ]] && record="${record%.*}"
 
 cli_token=$(cat /home/lauri/.secrets/linode/cli.token)
+domain_id=$(podman compose run --rm -e LINODE_CLI_TOKEN=$cli_token linode-cli \
+    domains ls | grep "\s$domain\s" || :)
 
-podman compose run --rm -e LINODE_CLI_TOKEN="$cli_token" linode-cli \
-    domains records-delete "$domain_id" "$record"
+record_id=
+
+if [[ "$domain_id" =~ [0-9]+ ]]; then
+    domain_id="${BASH_REMATCH[0]}"
+    record_id=$(podman compose run --rm -e LINODE_CLI_TOKEN=$cli_token linode-cli \
+	domains records-list "$domain_id" | grep "\s$record\s" || :)
+else
+    echo "Domainin hakeminen Linodelta epäonnistui"
+    exit 1
+fi
+
+if [[ "$record_id" =~ [0-9]+ ]]; then
+    podman compose run --rm -e LINODE_CLI_TOKEN="$cli_token" linode-cli \
+	domains records-delete "$domain_id" "$record_id"
+else
+    echo "Recordin hakeminen Linodelta epäonnistui"
+fi
 
 podman exec nginx nginx -s reload
